pipeline/requirements.txt
anthropic>=0.39.0
openai>=1.50.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
